--- .env.Example
@@ -1,4 +1,0 @@
-DB_NAME = ''
-DB_USER = ''
-DB_PASSWORD = ''
-SESSION_SECRET = ''
