+++ models/Comment.js
@@ -1,8 +1,29 @@
+// Desc: This file will represent the Comment model
+// ****Note: Likes and Dislikes and be considered as a type of vote
+// ==========================================================
+
+// Dependencies
+// =============================================================
+// Import model, datatypes, and Op from sequelize
+// Op is used to perform operations other than standard equals comparisons
 const { Model, DataTypes } = require('sequelize');
 const sequelize = require('../config/connection');
+// =============================================================
 
+// Create the Comment model, with methods
+// =============================================================
+// This Section will be used in further development
+//to add and remove likes and dislikes
+//to find comments with a minimum number of likes
+//to find comments with a minimum number of dislikes
+//to find comments with a maximum number of likes
+//to find comments with a maximum number of dislikes
+// =============================================================
 class Comment extends Model {}
+// =============================================================
 
+// Initialize the Comment model
+// =============================================================
 Comment.init(
   {
     id: {
@@ -38,5 +59,7 @@ Comment.init(
     modelName: 'comment',
   }
 );
+// =============================================================
 
+// Export the Comment model
 module.exports = Comment;
