--- models/Likes.js
@@ -1,31 +1,0 @@
-const { Model, DataTypes } = require('sequelize');
-const sequelize = require('../config/connection');
-
-class Likes extends Model {}
-
-Likes.init(
-  {
-    id: {
-      type: DataTypes.INTEGER,
-      allowNull: false,
-      primaryKey: true,
-      autoIncrement: true,
-    },
-    user_id:{
-      type: DataTypes.INTEGER,
-      allowNull: false,
-    },
-    comment_id:{
-      type: DataTypes.INTEGER,
-      allowNull: false,
-    },
-  },
-  {
-    sequelize,
-    freezeTableName: true,
-    underscored: true,
-    modelName: 'likes',
-  }
-);
-
-module.exports = Likes;
