+++ seeds/index.js
@@ -4,8 +4,6 @@
 // Importing the seeds files
 // =============================================================
 const sequelize = require('../config/connection');
-const seedComments = require('./comment-seeds');
-const seedMessages = require('./message-seeds');
 const seedItineraryItems = require('./itineraryitem-seeds');
 const seedTrip = require('./trip-seeds');
 const seedUsers = require('./user-seeds');
@@ -13,8 +11,6 @@ const seedTripSections = require('./tripsection-seeds');
 const seedTopic = require('./topicData');
 const seedPost = require('./postData');
 const seedComment = require('./commentData');
-const seedUser = require('./userData');
-const seedLikes = require('./likesData');
 // =============================================================
 
 // Syncing the seeds files
@@ -30,19 +26,12 @@ const seedAll = async () => {
     console.log('\n----- TRIP SECTIONS SEEDED -----\n');
     await seedItineraryItems();
     console.log('\n----- ITINERARY ITEMS SEEDED -----\n');
-    await seedMessages();
-    console.log('\n----- MESSAGES SEEDED -----\n');
-    await seedComments();
-    console.log('\n----- COMMENTS SEEDED -----\n');
-    await seedUser();
-
     await seedTopic();
-  
+    console.log('\n----- TOPICS SEEDED -----\n');  
     await seedPost();
-  
+    console.log('\n----- POSTS SEEDED -----\n');  
     await seedComment();
-  
-    await seedLikes();
+    console.log('\n----- COMMENTS SEEDED -----\n');  
     process.exit(0);
 };
